requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
aiohttp
aioredis>=2.0.0
redis
celery>=5.3
django-celery-beat
python-dotenv
setuptools
